+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Add extension for `Data` to track file I/O operations with Sentry (#4862)
+
 ## 8.45.0
 
 ### Features